# openapi.yml
openapi: 3.0.4
info:
  title: FinOrganizer API
  description: 'Personal finance tracking, portfolio, and planning API.'
  version: v1
paths:
...
  /api/v1/dashboard/net-worth:
    get:
      tags:
        - Dashboard
      parameters:
        - name: months
          in: query
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/NetWorthPointDto'
...
  /api/v1/dashboard/income-vs-expense:
    get:
      tags:
        - Dashboard
      parameters:
        - name: months
          in: query
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/IncomeVsExpenseMonthDto'
  /api/v1/dashboard/passive-income:
    get:
      tags:
        - Dashboard
      parameters:
        - name: months
          in: query
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PassiveIncomeSummaryDto'
  /api/v1/dashboard/top-expense-categories:
    get:
      tags:
        - Dashboard
      parameters:
        - name: from
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: to
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: top
          in: query
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TopExpenseCategoryDto'
  /api/v1/dashboard/projection:
    get:
      tags:
        - Dashboard
      parameters:
        - name: monthsAhead
          in: query
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProjectionMonthDto'
  /api/v1/savings-goals:
    get:
      tags:
        - Goals
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SavingsGoalDto'
    post:
      tags:
        - Goals
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSavingsGoalRequest'
        required: true
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SavingsGoalDto'
        '400':
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/HttpValidationProblemDetails'
  '/api/v1/savings-goals/{id}':
    get:
      tags:
        - Goals
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SavingsGoalDto'
        '404':
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
    put:
      tags:
        - Goals
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSavingsGoalRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SavingsGoalDto'
        '400':
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/HttpValidationProblemDetails'
        '404':
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
    delete:
      tags:
        - Goals
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: No Content
        '404':
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  '/api/v1/savings-goals/{id}/contributions':
    get:
      tags:
        - Goals
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SavingsGoalContributionDto'
    post:
      tags:
        - Goals
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddSavingsGoalContributionRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SavingsGoalContributionDto'
        '400':
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/HttpValidationProblemDetails'
        '404':
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
...
  /api/v1/transactions:
    get:
      tags:
        - Transactions
      parameters:
        - name: accountId
          in: query
          schema:
            type: string
            format: uuid
        - name: categoryId
          in: query
          schema:
            type: string
            format: uuid
        - name: type
          in: query
          schema:
            $ref: '#/components/schemas/TransactionType'
        - name: dateFrom
          in: query
          schema:
            type: string
            format: date
        - name: dateTo
          in: query
          schema:
            type: string
            format: date
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            format: int32
        - name: pageSize
          in: query
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionDtoPagedResult'
    post:
      tags:
        - Transactions
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTransactionRequest'
        required: true
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionDto'
        '400':
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/HttpValidationProblemDetails'
  '/api/v1/transactions/{id}':
    get:
      tags:
        - Transactions
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionDto'
        '404':
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
    put:
      tags:
        - Transactions
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTransactionRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionDto'
        '400':
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/HttpValidationProblemDetails'
        '404':
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
    delete:
      tags:
        - Transactions
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: No Content
        '404':
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /api/v1/transactions/import:
    post:
      tags:
        - Transactions
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ImportTransactionsRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionImportSummary'
components:
  schemas:
...
    AddSavingsGoalContributionRequest:
      type: object
      properties:
        amount:
          type: number
          format: double
        date:
          type: string
          format: date
        note:
          type: string
          nullable: true
      additionalProperties: false
...
    CreateSavingsGoalRequest:
      type: object
      properties:
        name:
          type: string
          nullable: true
        targetAmount:
          type: number
          format: double
        targetDate:
          type: string
          format: date
        linkedAccountId:
          type: string
          format: uuid
          nullable: true
      additionalProperties: false
    CreateTransactionRequest:
      type: object
      properties:
        accountId:
          type: string
          format: uuid
        type:
          $ref: '#/components/schemas/TransactionType'
        amount:
          type: number
          format: double
        currency:
          type: string
          nullable: true
        date:
          type: string
          format: date
        description:
          type: string
          nullable: true
        categoryId:
          type: string
          format: uuid
          nullable: true
        counterpartyAccountId:
          type: string
          format: uuid
          nullable: true
        tags:
          type: array
          items:
            type: string
          nullable: true
      additionalProperties: false
    CsvColumnMapping:
      type: object
      properties:
        dateColumn:
          type: integer
          format: int32
        amountColumn:
          type: integer
          format: int32
        descriptionColumn:
          type: integer
          format: int32
          nullable: true
        typeColumn:
          type: integer
          format: int32
          nullable: true
        categoryColumn:
          type: integer
          format: int32
          nullable: true
        tagsColumn:
          type: integer
          format: int32
          nullable: true
      additionalProperties: false
    CsvImportRowResult:
      type: object
      properties:
        rowNumber:
          type: integer
          format: int32
        transaction:
          $ref: '#/components/schemas/ParsedTransaction'
        error:
          type: string
          nullable: true
      additionalProperties: false
...
    HttpValidationProblemDetails:
      type: object
      properties:
        type:
          type: string
          nullable: true
        title:
          type: string
          nullable: true
        status:
          type: integer
          format: int32
          nullable: true
        detail:
          type: string
          nullable: true
        instance:
          type: string
          nullable: true
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
          nullable: true
      additionalProperties: { }
    ImportTransactionsRequest:
      type: object
      properties:
        csvContent:
          type: string
          nullable: true
        accountId:
          type: string
          format: uuid
        mapping:
          $ref: '#/components/schemas/CsvColumnMapping'
        hasHeaderRow:
          type: boolean
        delimiter:
          type: string
        dateFormat:
          type: string
          nullable: true
        currency:
          type: string
          nullable: true
        tagsSeparator:
          type: string
      additionalProperties: false
    IncomeVsExpenseMonthDto:
      type: object
      properties:
        month:
          type: string
          format: date
        income:
          type: number
          format: double
        expense:
          type: number
          format: double
        net:
          type: number
          format: double
      additionalProperties: false
    NetWorthPointDto:
      type: object
      properties:
        month:
          type: string
          format: date
        accountsBalance:
          type: number
          format: double
        portfolioValue:
          type: number
          format: double
        netWorth:
          type: number
          format: double
      additionalProperties: false
    ParsedTransaction:
      type: object
      properties:
        type:
          $ref: '#/components/schemas/TransactionType'
        amount:
          type: number
          format: double
        date:
          type: string
          format: date
        description:
          type: string
          nullable: true
        categoryName:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
          nullable: true
      additionalProperties: false
    PassiveIncomeMonthDto:
      type: object
      properties:
        month:
          type: string
          format: date
        amount:
          type: number
          format: double
      additionalProperties: false
    PassiveIncomeSummaryDto:
      type: object
      properties:
        months:
          type: array
          items:
            $ref: '#/components/schemas/PassiveIncomeMonthDto'
          nullable: true
        trailingTwelveMonthTotal:
          type: number
          format: double
      additionalProperties: false
    ProblemDetails:
      type: object
      properties:
        type:
          type: string
          nullable: true
        title:
          type: string
          nullable: true
        status:
          type: integer
          format: int32
          nullable: true
        detail:
          type: string
          nullable: true
        instance:
          type: string
          nullable: true
      additionalProperties: { }
    ProjectionAccountBalanceDto:
      type: object
      properties:
        accountId:
          type: string
          format: uuid
        accountName:
          type: string
          nullable: true
        projectedBalance:
          type: number
          format: double
      additionalProperties: false
    ProjectionMonthDto:
      type: object
      properties:
        month:
          type: string
          format: date
        recurringIncome:
          type: number
          format: double
        recurringExpense:
          type: number
          format: double
        net:
          type: number
          format: double
        cumulativeNetWorth:
          type: number
          format: double
        accountBalances:
          type: array
          items:
            $ref: '#/components/schemas/ProjectionAccountBalanceDto'
          nullable: true
      additionalProperties: false
...
    SavingsGoalContributionDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        savingsGoalId:
          type: string
          format: uuid
        amount:
          type: number
          format: double
        date:
          type: string
          format: date
        note:
          type: string
          nullable: true
      additionalProperties: false
    SavingsGoalDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          nullable: true
        targetAmount:
          type: number
          format: double
        targetDate:
          type: string
          format: date
        linkedAccountId:
          type: string
          format: uuid
          nullable: true
        currentAmount:
          type: number
          format: double
        progressPercent:
          type: number
          format: double
      additionalProperties: false
    TopExpenseCategoryDto:
      type: object
      properties:
        categoryId:
          type: string
          format: uuid
        categoryName:
          type: string
          nullable: true
        amount:
          type: number
          format: double
        percent:
          type: number
          format: double
      additionalProperties: false
    TransactionDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        accountId:
          type: string
          format: uuid
        type:
          $ref: '#/components/schemas/TransactionType'
        amount:
          type: number
          format: double
        currency:
          type: string
          nullable: true
        date:
          type: string
          format: date
        description:
          type: string
          nullable: true
        categoryId:
          type: string
          format: uuid
          nullable: true
        counterpartyAccountId:
          type: string
          format: uuid
          nullable: true
        recurrenceId:
          type: string
          format: uuid
          nullable: true
        tags:
          type: array
          items:
            type: string
          nullable: true
      additionalProperties: false
    TransactionDtoPagedResult:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/TransactionDto'
          nullable: true
        totalCount:
          type: integer
          format: int32
        page:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        totalPages:
          type: integer
          format: int32
          readOnly: true
      additionalProperties: false
    TransactionImportSummary:
      type: object
      properties:
        imported:
          type: integer
          format: int32
        errors:
          type: array
          items:
            $ref: '#/components/schemas/CsvImportRowResult'
          nullable: true
      additionalProperties: false
    TransactionType:
      enum:
        - 0
        - 1
        - 2
      type: integer
      format: int32
...
    UpdateSavingsGoalRequest:
      type: object
      properties:
        name:
          type: string
          nullable: true
        targetAmount:
          type: number
          format: double
        targetDate:
          type: string
          format: date
        linkedAccountId:
          type: string
          format: uuid
          nullable: true
      additionalProperties: false
    UpdateTransactionRequest:
      type: object
      properties:
        accountId:
          type: string
          format: uuid
        type:
          $ref: '#/components/schemas/TransactionType'
        amount:
          type: number
          format: double
        currency:
          type: string
          nullable: true
        date:
          type: string
          format: date
        description:
          type: string
          nullable: true
        categoryId:
          type: string
          format: uuid
          nullable: true
        counterpartyAccountId:
          type: string
          format: uuid
          nullable: true
        tags:
          type: array
          items:
            type: string
          nullable: true
      additionalProperties: false
...